添加回复

+++ src/main/java/life/majiang/community/controller/PublishController.java
@@ -23,7 +23,7 @@ public class PublishController {
     @Autowired
     QuestionService questionService;
     @GetMapping("/publish/{id}")
-    public String edit(@PathVariable("id") Integer id,
+    public String edit(@PathVariable("id") Long id,
                        Model model){
 
         QuestionDTO question = questionService.getById(id);
@@ -43,7 +43,7 @@ public class PublishController {
     public String doPublish(@RequestParam(value = "title",required = false) String title,
                             @RequestParam(value = "description",required = false) String description,
                             @RequestParam(value = "tag",required = false) String tag,
-                            @RequestParam(value = "id",required = false) Integer id,
+                            @RequestParam(value = "id",required = false) Long id,
                             HttpServletRequest request,
                             Model model
     ) {
